Add method to return the directory associated with a study version
Given a study version and accession, check the ftp server to obtain
the path to that study version directory. It is necessary to check
the ftp server instead of just constructing it since the directory
has the participant set number, which is not known from just the
study accession and version.